Refactor: reorganize simulation-related code
